added gradle step for executing UI-Test of generated testerra report in build.gradle; updated README.md; moved UI-Test to packed report_test; added TestReport.xml; updated AbstractReportTest;
