+++ report-ng-tests/README.md
@@ -1,10 +1,18 @@
 # ReportNG tests
 
-## Run 'Report under Test' tests
+## Run pretest
+Generate Testerra Report with several status
 
 ```shell
 gradle test -P pretest=true
 ```
+
+## Start Test of generated Testerra Report
+UI-Test of formerly generated Testerra Report
+
+```shell script
+gradle test
+````
 
 Or run all
 ```shell
