+++ report-ng-tests/build.gradle
@@ -10,24 +10,18 @@ dependencies {
 }
 
 test {
-    String[] s = []
+    String[] s = [
+        'src/test/resources/testng/TestReport.xml',
+    ]
 
 
     def preTest = findProperty("pretest")
     if (preTest) {
         s = ['src/test/resources/testng/ReportTestsUnderTest1.xml']
         System.setProperty("tt.report.dir", "src/test/resources/reports/report-generated-status")
-        System.setProperty("tt.report.name", "Generated Report for Test Status Check")
+        System.setProperty("tt.report.name", "Generated Report for UI-Test")
 
     }
-
-//    def nr = findProperty("nr")
-//    if (nr) {
-//        setFailureCorridor(nr)
-//        setCloseWindows()
-//        setReportDir("src/test/resources/reports/report-testsundertest-0" + nr)
-//        s = ['src/test/resources/testng/ReportTestsUnderTest' + nr + '.xml']
-//    }
 
     useTestNG() {
         suites s
@@ -38,48 +32,7 @@ test {
     }
 }
 
-//def setReportDir(reportname) {
-//    def reportdir = reportname
-//    def modulename = "report"
-//    if (!projectDir.toString().contains(modulename)) {
-//        reportdir = modulename + "/" + reportname
-//    }
-//    System.setProperty("tt.report.dir", reportdir)
-//}
 
-//def setFailureCorridor(nr) {
-//    System.setProperty("tt.failure.corridor.active", "true")
-//    System.setProperty("tt.failure.corridor.allowed.failed.tests.mid", "3")
-//    System.setProperty("tt.failure.corridor.allowed.failed.tests.low", "5")
-//    //abhängig von nr
-//    switch (nr) {
-//        case "1":
-//            System.setProperty("tt.failure.corridor.allowed.failed.tests.high", "34")
-//            break
-//        case "2":
-//        case "4":
-//        case "5":
-//            System.setProperty("tt.failure.corridor.allowed.failed.tests.high", "19")
-//            break
-//        case "3":
-//            System.setProperty("tt.failure.corridor.allowed.failed.tests.high", "24")
-//            break
-//        case "6":
-//            System.setProperty("tt.failure.corridor.allowed.failed.tests.high", "26")
-//            break
-//        case "7":
-//        case "8":
-//        case "9":
-//            System.setProperty("tt.failure.corridor.allowed.failed.tests.high", "0")
-//            System.setProperty("tt.failure.corridor.allowed.failed.tests.mid", "0")
-//            System.setProperty("tt.failure.corridor.allowed.failed.tests.low", "0")
-//            break
-//    }
-//}
-//
-//def setCloseWindows() {
-//    System.setProperty("tt.wdm.closeWindows.afterTestMethods", "false")
-//}
 
 task cleanReports {
     doLast {
