+++ report-ng-tests/src/test/resources/testng/TestReport.xml
@@ -1,0 +1,10 @@
+<?xml version="1.0" encoding="UTF-8"?>
+<!DOCTYPE suite SYSTEM "https://testng.org/testng-1.0.dtd">
+<suite name="Test Testerra Report" verbose="1" parallel="methods" configfailurepolicy="continue" thread-count="10" data-provider-thread-count="10">
+    <test name="Parallel" parallel="methods">
+        <packages>
+            <package name="io.testerra.report.test.report_test.methodpages"/>
+            <package name="io.testerra.report.test.report_test.sidebarpages"/>
+        </packages>
+    </test>
+</suite>
